microservices/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "Authentication and Authorization microservice"
authors = ["Dilip Kumar Sharma <diliplakshya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.test.dependencies]
httpx = "^0.23.3"
pytest = "^7.2.1"



[tool.poetry.group.docs.dependencies]
sphinx = "^6.1.3"


[tool.poetry.group.base.dependencies]
fastapi = {extras = ["all"], version = "^0.89.1"}
python-dotenv = "^0.21.1"


[tool.poetry.group.dev.dependencies]
sqlalchemy = "^2.0.1"
pymysql = "^1.0.2"
python-multipart = "^0.0.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^1.10.4"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
